% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aster.R
\name{aster}
\alias{aster}
\title{Add the significance asterisks}
\usage{
aster(x)
}
\arguments{
\item{x}{a \code{numeric} vector}
}
\value{
A character vector of \code{length(x)}
}
\description{
\code{aster} returns a character vector with the standard significance asterisks
}
\details{
Simply adds the asterisks "***", "**", "*" & "." for values below 0.001, 0.01, 0.05 & 0.1 respectively.

Doesn't add any legend, so this is just a quick convenience function.
}
\examples{
set.seed(42)
p <- rbeta(10, 1, 10)
data.frame(p = p, Sig = aster(p))
}
\seealso{
\code{\link{symnum}}
}

