% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capwords.R
\name{capwords}
\alias{capwords}
\title{Capitalise the first letter of every word}
\usage{
capwords(s, strict = TRUE, sep = " ", collapse = " ")
}
\arguments{
\item{s}{a character vector}

\item{strict}{logical enforces the lower case for all letters not beginning a word}

\item{collapse}{the character to separate the strings in the output}

\item{any}{character separating the words in the input \code{s}}
}
\value{
a character vector
}
\description{
Capitalise the first letter of every word
}
\details{
Taken from the examples on the page \code{\link{toupper}} and modified slightly.
}
\examples{
s <- c("a cat", "the_dog", "my bIRd")
capwords(s)
capwords(s, collapse = "_")
capwords(s, strict = FALSE)
capwords(s, sep="_", collapse=":")
}
\seealso{
\code{\link{toupper}} \code{\link{substring}}
}

