% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/commas.R
\name{commas}
\alias{commas}
\title{Print large values using commas}
\usage{
commas(x, ..., digits = 0, big.mark = ",")
}
\arguments{
\item{x}{a numeric vector}
}
\value{
A character vector of \code{length(x)}
}
\description{
Print large values using commas
}
\details{
A wrapper around the function \code{format}
}
\examples{
x <- pi*1e6
commas(x)
}
\seealso{
\code{\link{format}}
}

