% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrep.R
\name{magrep}
\alias{magrep}
\alias{magrepl}
\alias{magsub}
\title{Wrappers to grep and gsub functions}
\usage{
magrep(x, pattern, ...)

magrepl(x, pattern, ...)

magsub(x, pattern, replacement, ...)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which can be coerced by as.character to a character vector}

\item{pattern}{character string containing a \code{\link{regular expression}} (or character string for fixed = TRUE) 
to be matched in the given character vector.}

\item{...}{additional arguments passed to \code{\link{grep}}, \code{\link{grepl}} and \code{\link{gsub}}}

\item{replacement}{a replacement for matched pattern in sub and gsub. Coerced to character if possible.}
}
\description{
Wrappers to common grep functions for compatability with the magrittr style of programming
}
\details{
These return the same as the functions \code{\link{grep}}, \code{\link{grepl}} and \code{\link{gsub}}, 
but have the character placed first for easy integration with the magrittr pipe and lapply functions.
}
\seealso{
\code{\link{grep}}, \code{\link{grepl}} and \code{\link{gsub}}
}

