% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc.R
\name{uc}
\alias{uc}
\title{Return the upper left corner of a matrix or data frame}
\usage{
uc(x, nr = 6, nc, ...)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame}}

\item{nr}{the number of rows to return}

\item{nc}{the number of columns to return}
}
\value{
an object of class(x)
}
\description{
Return the upper left corner of a matrix or data.frame
}
\details{
Just used for quick visual inspection of large R objects as an alternative to \code{head}
}
\examples{
x <- matrix(rnorm(1000), nrow = 10)
uc(x)

}
\seealso{
\code{\link{head}}
}

