% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitLine.R
\name{unitLine}
\alias{unitLine}
\title{Add the unit line to a plot}
\usage{
unitLine(col = "blue", ...)
}
\arguments{
\item{...}{used to pass any parameters to \code{\link{abline}}}
}
\description{
Adds the unit line to plot
}
\details{
This is a convenience wrapper to the function \code{\link{abline}} from the default graphics package
}
\examples{
set.seed(31)
x <- rnorm(10)
y <- rnorm(10)
plot(x,y)
unitLine()

}

