% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uc.R
\name{uc}
\alias{uc}
\title{Return the upper left corner of a matrix}
\usage{
uc(x, nr = 6, nc = 6, ...)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame}}

\item{nr}{the number of rows to return}

\item{nc}{the number of columns to return}
}
\value{
an object of class(x)
}
\description{
Return the upper left corner of a matrix
}
\details{
Just used for quick visual inspection of large matrices as an alternative to \code{head}
}
\examples{
x <- matrix(rnorm(1000), nrow = 10)
uc(x)
}
\seealso{
\code{\link{head}}
}

